+++ src/solvers/difference-of-convex-proximal-point.jl
@@ -1,6 +1,7 @@
 
 @doc raw"""
-    DifferenceOfConvexProximalState{Type} <: Options
+    DifferenceOfConvexProximalState{Type} <: AbstractSubProblemSolverState
+
 A struct to store the current state of the algorithm as well as the form.
 It comes in two forms, depending on the realisation of the `subproblem`.
 
@@ -14,6 +15,8 @@ It comes in two forms, depending on the realisation of the `subproblem`.
 * `stop`:                      ([`StopWhenChangeLess`](@ref)`(1e-8)`) a [`StoppingCriterion`](@ref)
 * `X`, `Y`:                    (`zero_vector(M,p)`) the current gradient and descent direction, respectively
   their common type is set by the keyword `X`
+* `sub_problem`:               an [`AbstractManoptProblem`](@ref) problem or a function `(M, p, X) -> q` or `(M, q, p, X)` for the a closed form solution of the sub problem
+* `sub_state`:                 an [`AbstractManoptSolverState`](@ref) for the subsolver or an [`AbstractEvaluationType`](@ref) in case the sub problem is provided as a function
 
 # Constructor
 
@@ -28,7 +31,7 @@ mutable struct DifferenceOfConvexProximalState{
     P,
     T,
     Pr,
-    St,
+    St<:AbstractManoptSolverState,
     S<:Stepsize,
     SC<:StoppingCriterion,
     RTR<:AbstractRetractionMethod,
@@ -50,7 +53,7 @@ mutable struct DifferenceOfConvexProximalState{
         M::AbstractManifold,
         p::P,
         sub_problem::Pr,
-        sub_state::St;
+        sub_state::Union{AbstractEvaluationType,AbstractManoptSolverState};
         X::T=zero_vector(M, p),
         stepsize::S=ConstantStepsize(M),
         stopping_criterion::SC=StopWhenChangeLess(M, 1e-8),
@@ -61,20 +64,20 @@ mutable struct DifferenceOfConvexProximalState{
         P,
         T,
         Pr,
-        St,
         S<:Stepsize,
         SC<:StoppingCriterion,
         I<:AbstractInverseRetractionMethod,
         R<:AbstractRetractionMethod,
         Fλ,
     }
-        return new{P,T,Pr,St,S,SC,R,I,Fλ}(
+        sub_state_storage = maybe_wrap_evaluation_type(sub_state)
+        return new{P,T,Pr,typeof(sub_state_storage),S,SC,R,I,Fλ}(
             λ,
             p,
             copy(M, p),
             copy(M, p),
             sub_problem,
-            sub_state,
+            sub_state_storage,
             X,
             retraction_method,
             inverse_retraction_method,
@@ -358,7 +361,7 @@ function difference_of_convex_proximal_point!(
     sub_state::Union{AbstractEvaluationType,AbstractManoptSolverState,Nothing}=if !isnothing(
         prox_g
     )
-        evaluation
+        maybe_wrap_evaluation_type(evaluation)
     elseif isnothing(sub_objective)
         nothing
     else
@@ -404,7 +407,7 @@ function difference_of_convex_proximal_point!(
         M,
         p,
         sub_problem,
-        sub_state;
+        maybe_wrap_evaluation_type(sub_state);
         X=X,
         stepsize=stepsize,
         stopping_criterion=stopping_criterion,
@@ -425,7 +428,9 @@ end
 =#
 function step_solver!(
     amp::AbstractManoptProblem,
-    dcps::DifferenceOfConvexProximalState{P,T,<:Function,AllocatingEvaluation},
+    dcps::DifferenceOfConvexProximalState{
+        P,T,<:Function,ClosedFormSubSolverState{AllocatingEvaluation}
+    },
     i,
 ) where {P,T}
     M = get_manifold(amp)
@@ -444,7 +449,9 @@ end
 =#
 function step_solver!(
     amp::AbstractManoptProblem,
-    dcps::DifferenceOfConvexProximalState{P,T,<:Function,InplaceEvaluation},
+    dcps::DifferenceOfConvexProximalState{
+        P,T,<:Function,ClosedFormSubSolverState{InplaceEvaluation}
+    },
     i,
 ) where {P,T}
     M = get_manifold(amp)