+++ src/solvers/exact_penalty_method.jl
@@ -27,8 +27,13 @@ construct an exact penalty options with the remaining previously mentioned field
 
 [`exact_penalty_method`](@ref)
 """
-mutable struct ExactPenaltyMethodState{P,Pr,St,R<:Real,TStopping<:StoppingCriterion} <:
-               AbstractSubProblemSolverState
+mutable struct ExactPenaltyMethodState{
+    P,
+    Pr<:Union{F,AbstractManoptProblem} where {F},
+    St<:AbstractManoptSolverState,
+    R<:Real,
+    TStopping<:StoppingCriterion,
+} <: AbstractSubProblemSolverState
     p::P
     sub_problem::Pr
     sub_state::St
@@ -45,7 +50,7 @@ mutable struct ExactPenaltyMethodState{P,Pr,St,R<:Real,TStopping<:StoppingCriter
         M::AbstractManifold,
         p::P,
         sub_problem::Pr,
-        sub_state::St;
+        sub_state::Union{AbstractEvaluationType,AbstractManoptSolverState};
         ϵ::R=1e-3,
         ϵ_min::R=1e-6,
         ϵ_exponent=1 / 100,
@@ -59,17 +64,12 @@ mutable struct ExactPenaltyMethodState{P,Pr,St,R<:Real,TStopping<:StoppingCriter
         stopping_criterion::SC=StopAfterIteration(300) | (
             StopWhenSmallerOrEqual(:ϵ, ϵ_min) | StopWhenChangeLess(M, 1e-10)
         ),
-    ) where {
-        P,
-        Pr<:AbstractManoptProblem,
-        St<:AbstractManoptSolverState,
-        R<:Real,
-        SC<:StoppingCriterion,
-    }
-        epms = new{P,Pr,St,R,SC}()
+    ) where {P,Pr<:Union{F,AbstractManoptProblem} where {F},R<:Real,SC<:StoppingCriterion}
+        sub_state_storage = maybe_wrap_evaluation_type(sub_state)
+        epms = new{P,Pr,typeof(sub_state_storage),R,SC}()
         epms.p = p
         epms.sub_problem = sub_problem
-        epms.sub_state = sub_state
+        epms.sub_state = sub_state_storage
         epms.ϵ = ϵ
         epms.ϵ_min = ϵ_min
         epms.u = u
@@ -240,16 +240,6 @@ function exact_penalty_method(
     grad_g_ = _ensure_mutating_gradient(grad_g, p, evaluation)
     h_ = _ensure_mutating_cost(h, p)
     grad_h_ = _ensure_mutating_gradient(grad_h, p, evaluation)
-    num_eq = if isnothing(equality_constrains)
-        _number_of_constraints(h_, grad_h_; M=M, p=p_)
-    else
-        inequality_constrains
-    end
-    num_ineq = if isnothing(inequality_constrains)
-        _number_of_constraints(g_, grad_g_; M=M, p=p_)
-    else
-        inequality_constrains
-    end
     cmo = ConstrainedManifoldObjective(
         f_,
         grad_f_,
@@ -258,8 +248,8 @@ function exact_penalty_method(
         h_,
         grad_h_;
         evaluation=evaluation,
-        equality_constraints=num_eq,
-        inequality_constraints=num_ineq,
+        equality_constraints=equality_constraints,
+        inequality_constraints=equality_constraints,
         M=M,
         p=p_,
     )
@@ -356,7 +346,7 @@ function exact_penalty_method!(
     sub_cost=ExactPenaltyCost(cmo, ρ, u; smoothing=smoothing),
     sub_grad=ExactPenaltyGrad(cmo, ρ, u; smoothing=smoothing),
     sub_kwargs=(;),
-    sub_problem::AbstractManoptProblem=DefaultManoptProblem(
+    sub_problem::Pr=DefaultManoptProblem(
         M,
         decorate_objective!(
             M,
@@ -368,7 +358,7 @@ function exact_penalty_method!(
     sub_stopping_criterion=StopAfterIteration(300) |
                            StopWhenGradientNormLess(ϵ) |
                            StopWhenStepsizeLess(1e-8),
-    sub_state::AbstractManoptSolverState=decorate_state!(
+    sub_state::Union{AbstractEvaluationType,AbstractManoptSolverState}=decorate_state!(
         QuasiNewtonState(
             M,
             copy(M, p);
@@ -386,12 +376,16 @@ function exact_penalty_method!(
         StopWhenSmallerOrEqual(:ϵ, ϵ_min) & StopWhenChangeLess(M, 1e-10)
     ),
     kwargs...,
-) where {O<:Union{ConstrainedManifoldObjective,AbstractDecoratedManifoldObjective}}
+) where {
+    O<:Union{ConstrainedManifoldObjective,AbstractDecoratedManifoldObjective},
+    Pr<:Union{F,AbstractManoptProblem} where {F},
+}
+    sub_state_storage = maybe_wrap_evaluation_type(sub_state)
     emps = ExactPenaltyMethodState(
         M,
         p,
         sub_problem,
-        sub_state;
+        sub_state_storage;
         ϵ=ϵ,
         ϵ_min=ϵ_min,
         u=u,