+++ test/solvers/test_adaptive_regularization_with_cubics.jl
@@ -132,14 +132,18 @@ include("../utils/example_tasks.jl")
         f1(M, p) = p
         f1!(M, q, p) = copyto!(M, q, p)
         r = copy(M, p1)
-        Manopt.solve_arc_subproblem!(M, r, f1, AllocatingEvaluation(), p0)
+        Manopt.solve_arc_subproblem!(
+            M, r, f1, Manopt.ClosedFormSubSolverState{AllocatingEvaluation}(), p0
+        )
         @test r == p0
         r = copy(M, p1)
-        Manopt.solve_arc_subproblem!(M, r, f1!, InplaceEvaluation(), p0)
+        Manopt.solve_arc_subproblem!(
+            M, r, f1!, Manopt.ClosedFormSubSolverState{InplaceEvaluation}(), p0
+        )
         @test r == p0
         # Dummy construction with a function for the `sub_problem`
         arcs4 = AdaptiveRegularizationState(M, p0; sub_problem=f1)
-        @test arcs4.sub_state isa AbstractEvaluationType
+        @test arcs4.sub_state isa Manopt.ClosedFormSubSolverState
     end
 
     @testset "A few solver runs" begin